docs: build vs install done
go build just compile the executable file and move it to the destination. go install move the executable file to $GOPATH/bin if $GOBIN is set and cache all non-main packages which imported to $GOPATH/pkg. The cache will be used in the next compile if it not changed yet.